+++ packages/web3-eth-contract/test/integration/contract_accesslist.test.ts
@@ -22,10 +22,11 @@ import {
 	createTempAccount,
 	describeIf,
 	getSystemTestBackend,
+	BACKEND
 } from '../fixtures/system_test_utils';
 
 describe('contract', () => {
-	describeIf(getSystemTestBackend() === 'geth')('createAccessList', () => {
+	describeIf(getSystemTestBackend() === BACKEND.GETH)('createAccessList', () => {
 		let contract: Contract<typeof GreeterAbi>;
 		let deployOptions: Record<string, unknown>;
 		let sendOptions: Record<string, unknown>;