+++ packages/web3-core/src/web3_request_manager.ts
@@ -425,10 +425,10 @@ export class Web3RequestManager<
 		} else if ((response as unknown) instanceof Error) {
 			error = response as unknown as JsonRpcError;
 		}
-
+		
 		// This message means that there was an error while executing the code of the smart contract
 		// However, more processing will happen at a higher level to decode the error data,
-		//	according to the Error ABI, if it was available as of EIP-838.
+		//	according to the Error ABI, if it was available as of EIP-838. 
 		if (error?.message.includes('revert')) throw new ContractExecutionError(error);
 
 		return false;