+++ packages/web3-eth/test/integration/web3_eth/sign.test.ts
@@ -21,7 +21,9 @@ import { Web3Eth } from '../../../src';
 import {
 	closeOpenConnection,
 	createTempAccount,
+	getSystemTestBackend,
 	getSystemTestProvider,
+	BACKEND
 } from '../../fixtures/system_test_utils';
 
 describe('Web3Eth.sign', () => {
@@ -38,8 +40,17 @@ describe('Web3Eth.sign', () => {
 	});
 
 	it('should sign message', async () => {
-		const message = '0x736f796c656e7420677265656e2069732070656f706c65';
-		const response = await web3Eth.sign(message, tempAcc.address);
-		expect(isHexStrict(response as string)).toBe(true);
+		if (getSystemTestBackend() === BACKEND.HARDHAT) {
+			const message = '0x736f796c656e7420677265656e2069732070656f706c65';
+			const accounts = await web3Eth.getAccounts();
+			const response = await web3Eth.sign(message, accounts[0]);
+			// eslint-disable-next-line jest/no-conditional-expect
+			expect(isHexStrict(response as string)).toBe(true);
+		} else {
+			const message = '0x736f796c656e7420677265656e2069732070656f706c65';
+			const response = await web3Eth.sign(message, tempAcc.address);
+			// eslint-disable-next-line jest/no-conditional-expect
+			expect(isHexStrict(response as string)).toBe(true);
+		}
 	});
 });